added CheckForm component
